+++ Client/Chat.hpp
@@ -14,7 +14,7 @@ class Chat : public QObject
 {
     Q_OBJECT
 
-    Q_PROPERTY(int64_t messageId MEMBER message_id NOTIFY messageIdChanged)
+    Q_PROPERTY(int messageId MEMBER message_id NOTIFY messageIdChanged)
     Q_PROPERTY(QString sessionId MEMBER session_id NOTIFY sessionIdChanged)
     Q_PROPERTY(QString senderId MEMBER sender_id NOTIFY senderIdChanged)
     Q_PROPERTY(QString senderName MEMBER sender_name NOTIFY senderNameChanged)
@@ -30,7 +30,7 @@ class Chat : public QObject
         {TEXT_CHAT, "qrc:/qml/TextChat.qml"}};
 
   public:
-    int64_t message_id;
+    int message_id;
     QString session_id;
     QString sender_id;
     QString sender_name;
@@ -43,7 +43,7 @@ class Chat : public QObject
     std::set<QString> reader_set;
     bool is_oppoent;
 
-    Chat(const int64_t &message_id = 0,
+    Chat(const int &message_id = 0,
          const QString &session_id = "",
          const QString &sender_id = "",
          const QString &sender_name = "",
